Fix: Fix all the QA reviews
